backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-environ==0.8.1
django-rest-framework==0.1.0
django-simple-history==3.0.0
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
